Move Stmt into GL.Shader.DSL.
