+++ starlight.cabal
@@ -86,7 +86,6 @@ library
     GL.Shader.GLSL
     GL.Shader.GLSL.Syntax
     GL.Shader.GLSL.Type
-    GL.Shader.Stmt
     GL.Shader.Vars
     GL.Texture
     GL.TextureUnit
